Fix SCC evaluation order issue
Previously we executed the algorithm in phases across the whole graph.
However, this could lead to a redundant SCC not being found:
If a redundant SCC B refers only to different nodes within a redundant
SCC A, but A is nested more deeply in SCCs, then A would eventually be
found, but only after B has been discarded from the working set.

This has been fixed by evaluating the SCCs in the order intended by the
algorithm. Oops.